requirements.txt
aiohttp==3.7.4.post0
beautifulsoup4==4.9.3
discord.py @ git+https://github.com/Rapptz/discord.py.git@489e5f3288efb9b9e6649447209e1ac1746ccb32
jishaku==2.2.0
Pillow==8.3.1
requests-html==0.10.0
tabulate==0.8.9
typing-extensions==3.10.0.0
numpy==1.21.4
matplotlib
Equation